+++ domain/src/main/java/com/wanted/domain/repository/SearchRepository.kt
@@ -4,5 +4,5 @@ import com.wanted.domain.model.PagedCompanyModel
 import com.wanted.domain.result.DomainResult
 
 interface SearchRepository {
-    suspend fun getSearchCompany(query: String, offset: Int, limit: Long): DomainResult<PagedCompanyModel>
+    suspend fun getSearchCompany(query: String, offset: Int, limit: Int): DomainResult<PagedCompanyModel>
 }
